gallan: extract more stuff out of S3Archiver.
I've put the batch transformation, compression, etc all together in one place, and have it expose it's result as an InputStream. Heap usage is a real constraint when used for 100k batch sizes, I'd like to see this be able to stream protobufs + compression as part of the upload, to reduce heap usage.